# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TechLane API
  version: 1.0.0
  description: Shared REST API for TechLane ops web, Android, and portals.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /sync/commands:
    post:
      summary: Submit offline sync command
      operationId: submitSyncCommand
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncCommand"
      responses:
        "200":
          description: Applied or idempotent replay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncCommandResult"
...
  /notifications:
    get:
      summary: Staff notification inbox
      operationId: listNotifications
      parameters:
        - name: unacked
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Inbox items
  /notifications/{id}/ack:
    post:
      summary: Acknowledge staff notification
      operationId: ackNotification
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Acked
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema:
        type: string
        format: uuid
  schemas:
...
    SyncCommand:
      type: object
      required: [action_id, command_type, payload]
      properties:
        action_id:
          type: string
          format: uuid
        command_type:
          type: string
        branch_id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        local_timestamp:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
    SyncCommandResult:
      type: object
      properties:
        action_id:
          type: string
          format: uuid
        sync_status:
          type: string
        result:
          type: object
          additionalProperties: true
        error:
          type: string